separate allocator directory
