+++ src/allocator/tests/allocator/mod.rs
@@ -1,0 +1,3 @@
+//! Defines tests for [`crate::allocator::Allocator`] interface.
+
+mod allocate;
