+++ src/allocator/tests/mod.rs
@@ -1,0 +1,4 @@
+//! Defines tests for interfaces defined in [`crate::allocator`] module.
+
+mod allocator;
+mod slice;
