+++ src/allocator/tests/slice/mod.rs
@@ -1,0 +1,3 @@
+//! Defines tests for [`crate::allocator::slice::Slice`] interface.
+
+mod iter;
